+++ src/pages/admin/components/team/MemberProfile.tsx
@@ -1,6 +1,6 @@
 import Text from '@gdsc/components/common/typography/Text';
 
-import type { TeamMember } from '@gdsc/apis/hooks/admin/useGetTeamMember';
+import type { TeamMember } from '@gdsc/apis/hooks/admin/team/useGetTeamMember';
 
 import { MemberBox, TextWrapper } from './TeamBox.style';
 import { Draggable } from '@hello-pangea/dnd';
